Restructure rerun coalescing E2E coverage
